+++ src/main/kotlin/com/github/copyinaction/booking/domain/BookingSeat.kt
@@ -44,6 +44,8 @@ class BookingSeat(
 ) : BaseEntity() {
 
     companion object {
+        const val DEFAULT_SECTION = "GENERAL"
+
         fun create(
             booking: Booking,
             section: String,